changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11753 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/Views/ScriptView/ORLineNumberingRulerView.h
@@ -48,7 +48,7 @@
 - (NSColor*) backgroundColor;
 
 - (NSUInteger) lineNumberForLocation:(float)location;
-- (ORLineMarker*) markerAtLine:(unsigned)line;
+- (ORLineMarker*) markerAtLine:(NSUInteger)line;
 - (void) loadLineMarkers:(NSDictionary*)someLineMarkers;
 
 @end
